+++ src/i18n/legacy/ja.yml
