+++ src/i18n/legacy/locale.ts
